refactor(multipooler): move postgres monitor code into postgres_monitor.go
manager.go had grown to ~2470 lines. Move the PostgreSQL monitor / auto-recovery
code into its own file, matching the package's existing split-by-concern layout
(rpc_consensus.go, pg_replication.go, etc.). Pure code movement — functions and
types moved verbatim, no behavior change.

Moved: the rule-to-role helpers (deriveTypeAndObs, latestRule, intendedRole,
staleStandbyDemoteTarget), the monitor state types (postgresState,
remedialAction), the monitor loop body (monitorPostgresIteration,
discoverPostgresState), the decision functions (determineRemedialAction and its
determineRoleAction / determineReplicationSettingsAction /
determinePostgresNotRunningAction helpers), takeRemedialAction, and the
postgres bring-up helpers (hasCompleteBackups, startPostgres,
restoreAndStartPostgres). The monitor goroutine is still wired up from
openLocked in manager.go.
